requirements.txt
dlib
face_recognition
imutils
opencv-python
djangorestframework
django
django-cors-headers